pyproject.toml
[tool.poetry]
name = "postgre-dbt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dbt-core = "^1.8.6"
dbt-postgres = "^1.8.2"
pandas = "^2.2.2"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
